+++ src/main/java/koreatech/in/service/OwnerServiceImpl.java
@@ -2,7 +2,6 @@ package koreatech.in.service;
 
 import java.io.IOException;
 import java.sql.SQLException;
-import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import koreatech.in.domain.User.EmailAddress;
 import koreatech.in.domain.User.User;
@@ -13,7 +12,6 @@ import koreatech.in.domain.User.owner.OwnerAttachments;
 import koreatech.in.domain.User.owner.OwnerInCertification;
 import koreatech.in.domain.User.owner.OwnerInVerification;
 import koreatech.in.domain.User.owner.OwnerShop;
-import koreatech.in.dto.normal.mail.MailForm;
 import koreatech.in.dto.normal.user.owner.request.OwnerRegisterRequest;
 import koreatech.in.dto.normal.user.owner.request.OwnerUpdateRequest;
 import koreatech.in.dto.normal.user.owner.request.VerifyCodeRequest;
@@ -25,7 +23,6 @@ import koreatech.in.exception.ExceptionInformation;
 import koreatech.in.mapstruct.OwnerConverter;
 import koreatech.in.repository.user.OwnerMapper;
 import koreatech.in.repository.user.UserMapper;
-import koreatech.in.util.RandomGenerator;
 import koreatech.in.util.SesMailSender;
 import koreatech.in.util.SlackNotiSender;
 import koreatech.in.util.StringRedisUtilStr;
@@ -37,7 +34,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static koreatech.in.dto.normal.mail.MailForm.*;
+import static koreatech.in.domain.Mail.MailForm.*;
 
 @Service
 public class OwnerServiceImpl implements OwnerService {
